+++ packages/moonstone/Scroller/Scroller.js
@@ -165,9 +165,9 @@ class ScrollerBase extends Component {
 	/**
 	 * Returns the first spotlight container between `node` and the scroller
 	 *
-	 * @param   {Node}      node  A DOM node
+	 * @param {Node} node A DOM node
 	 *
-	 * @returns {Node|Null}       Spotlight container for `node`
+	 * @returns {Node|Null} Spotlight container for `node`
 	 * @private
 	 */
 	getSpotlightContainerForNode = (node) => {
@@ -182,9 +182,9 @@ class ScrollerBase extends Component {
 	 * Calculates the "focus bounds" of a node. If the node is within a spotlight container, that
 	 * container is scrolled into view rather than just the element.
 	 *
-	 * @param   {Node}   node  Focused node
+	 * @param {Node} node Focused node
 	 *
-	 * @returns {Object}       Bounds as returned by `getBoundingClientRect`
+	 * @returns {Object} Bounds as returned by `getBoundingClientRect`
 	 * @private
 	 */
 	getFocusedItemBounds = (node) => {
@@ -225,18 +225,30 @@ class ScrollerBase extends Component {
 		return oPoint;
 	}
 
-	calculatePositionOnFocus = (focusedItem, scrollInfo) => {
-		if (!this.isVertical() && !this.isHorizontal() || !focusedItem || !this.containerRef.contains(focusedItem)) return;
+	/**
+	 * Calculates the new top and left position for scroller based on focusedItem.
+	 *
+	 * @param {Node} item node
+	 * @param {Object} scrollInfo position info. `calculateScrollTop` uses
+	 * `scrollInfo.previousScrollHeight` and `scrollInfo.scrollTop`
+	 *
+	 * @returns {Object} with keys {top, left} containing caculated top and left positions for scroll.
+	 * @private
+	 */
+	calculatePositionOnFocus = ({item, scrollInfo}) => {
+		if (!this.isVertical() && !this.isHorizontal() || !item || !this.containerRef.contains(item)) {
+			return;
+		}
 
 		const {
 			top: itemTop,
 			left: itemLeft,
 			height: itemHeight,
 			width: itemWidth
-		} = this.getFocusedItemBounds(focusedItem);
+		} = this.getFocusedItemBounds(item);
 
 		if (this.isVertical()) {
-			this.scrollPos.top = this.calculateScrollTop(focusedItem, itemTop, itemHeight, scrollInfo);
+			this.scrollPos.top = this.calculateScrollTop(item, itemTop, itemHeight, scrollInfo);
 		}
 
 		if (this.isHorizontal()) {
@@ -262,6 +274,18 @@ class ScrollerBase extends Component {
 		return this.scrollPos;
 	}
 
+	/**
+	 * Calculates the new `scrollTop`.
+	 *
+	 * @param {Node} focusedItem node
+	 * @param {Number} itemTop of the focusedItem / focusedContainer
+	 * @param {Number} itemHeight of focusedItem / focusedContainer
+	 * @param {Object} scrollInfo position info. Uses `scrollInfo.previousScrollHeight`
+	 * and `scrollInfo.scrollTop`
+	 *
+	 * @returns {Number} Calculated `scrollTop`
+	 * @private
+	 */
 	calculateScrollTop = (focusedItem, itemTop, itemHeight, scrollInfo) => {
 		const
 			{clientHeight} = this.scrollBounds,
@@ -274,6 +298,7 @@ class ScrollerBase extends Component {
 
 		let newScrollTop = this.scrollPos.top;
 
+		// Caculations for when scrollHeight decrease.
 		if (scrollInfo) {
 			const
 				{scrollTop, previousScrollHeight} = scrollInfo,
@@ -283,7 +308,6 @@ class ScrollerBase extends Component {
 			newScrollTop = scrollTop;
 
 			if (scrollHeightDecrease > 0) {
-				// Update scrollTop for scrollHeight decrease
 				const
 					itemBounds = focusedItem.getBoundingClientRect(),
 					newItemBottom = newScrollTop + itemBounds.top + itemBounds.height - containerTop;
@@ -291,7 +315,8 @@ class ScrollerBase extends Component {
 				if (newItemBottom < scrollBottom && scrollHeightDecrease + newItemBottom > scrollBottom) {
 					// When `focusedItem` is not at the very bottom of the `Scroller` and
 					// `scrollHeightDecrease` caused a scroll.
-					const distanceFromBottom = scrollBottom - newItemBottom,
+					const
+						distanceFromBottom = scrollBottom - newItemBottom,
 						bottomOffset = scrollHeightDecrease - distanceFromBottom;
 					if (bottomOffset < newScrollTop) {
 						// guard against negative `scrollTop`
@@ -307,8 +332,8 @@ class ScrollerBase extends Component {
 			}
 		}
 
+		// Calculations for `containerHeight` that are bigger than `clientHeight`
 		if (itemHeight > clientHeight) {
-			// scroller behavior for containers that are bigger than `clientHeight`
 			const
 				{top, height: nestedItemHeight} = focusedItem.getBoundingClientRect(),
 				nestedItemTop = this.containerRef.scrollTop + (top - containerTop),
@@ -318,15 +343,20 @@ class ScrollerBase extends Component {
 				// set scroll position so that the top of the container is at least on the top
 				newScrollTop = newItemTop - nestedItemHeight;
 			} else if (nestedItemBottom > scrollBottom) {
+				// Caculate when 5-way focus down past the bottom.
 				newScrollTop += nestedItemBottom - scrollBottom;
 			} else if (nestedItemTop < currentScrollTop) {
+				// Caculate when 5-way focus up past the top.
 				newScrollTop += nestedItemTop - currentScrollTop;
 			}
 		} else if (itemBottom > scrollBottom) {
+			// Caculate when 5-way focus down past the bottom.
 			newScrollTop += itemBottom - scrollBottom;
 		} else if (newItemTop < currentScrollTop) {
+			// Caculate when 5-way focus up past the top.
 			newScrollTop += newItemTop - currentScrollTop;
 		}
+
 		return newScrollTop;
 	}