+++ packages/spotlight/CHANGELOG.md
@@ -10,7 +10,11 @@ The following is a curated list of changes in the Enact spotlight module, newest
 
 ### Changed
 
+- `spotlight` containers using a `restrict` value of 'self-only' will ignore `leaveFor` directives when attempting to leave the container via 5-way
+
 ### Fixed
+
+- `spotlight` to not blur and re-focus an element that is already focused
 
 ## [1.5.0] - 2017-07-19